Split diff models into several files
