+++ nbdime-web/src/diff/model/index.ts
@@ -1,0 +1,23 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+'use strict';
+
+export {
+  CellDiffModel
+} from './cell';
+
+export {
+  IDiffModel
+} from './common';
+
+export {
+  NotebookDiffModel
+} from './notebook';
+
+export {
+  OutputDiffModel
+} from './output';
+
+export {
+  IStringDiffModel, StringDiffModel
+} from './string';
